cheri: Correctly use uintptr_t in fs/, drivers/, ipc/ and init/
Correctly use uintptr_t and associated casts in fs/, drivers/
ipc/ and init/.

+++ drivers/tty/serial/serial_core.c
@@ -295,7 +295,7 @@ static void uart_free_xmit_buf(struct tty_port *port)
 	INIT_KFIFO(port->xmit_fifo);
 	uart_port_unlock(uport, flags);
 
-	free_page((unsigned long)xmit_buf);
+	free_page((uintptr_t)xmit_buf);
 }
 
 /*
