+++ drivers/tty/vt/vc_screen.c
@@ -695,7 +695,7 @@ vcs_write(struct file *file, const char __user *buf, size_t count, loff_t *ppos)
 		buf += this_round;
 		pos += this_round;
 		if (org)
-			update_region(vc, (unsigned long)(org0), org - org0);
+			update_region(vc, (uintptr_t)(org0), org - org0);
 	}
 	*ppos += written;
 	ret = written;
