+++ fs/proc/inode.c
@@ -373,7 +373,7 @@ static __poll_t proc_reg_poll(struct file *file, struct poll_table_struct *pts)
 	return rv;
 }
 
-static long pde_ioctl(struct proc_dir_entry *pde, struct file *file, unsigned int cmd, unsigned long arg)
+static long pde_ioctl(struct proc_dir_entry *pde, struct file *file, unsigned int cmd, uintptr_t arg)
 {
 	typeof_member(struct proc_ops, proc_ioctl) ioctl;
